+++ src/gromo/utils/tensor_statistic.py
@@ -21,16 +21,14 @@ class TensorStatistic:
         loader `data_loader`. We can use the following code:
 
             ```python
-            data = torch.zeros(2, 3)
             tensor_statistic = TensorStatistic(
                 shape=(2, 3),
-                update_function=lambda: (data.sum(dim=0), data.size(0)),
+                update_function=lambda data: (data.sum(dim=0), data.size(0)),
                 name="Average",
             )
             for data_batch in data_loader:
-                data = data_batch
                 tensor_statistic.updated = False
-                tensor_statistic.update()
+                tensor_statistic.update(data_batch)
 
             print(tensor_statistic())
             ```
@@ -73,32 +71,26 @@ class TensorStatistic:
         return f"{self.name} tensor of shape {self._shape} with {self.samples} samples"
 
     @torch.no_grad()
-    def update(self, **kwargs):
-        assert (
-            not self._shape or self._tensor is not None
-        ), f"The tensor statistic {self.name} has not been initialized."
+    def update(self, **kwargs) -> tuple[torch.Tensor, int] | None:
         if self.updated is False:
-            update, nb_sample = self._update_function(**kwargs)
-            if self._tensor is not None:
-                assert update.size() == self._tensor.size(), (
-                    f"The update tensor has a different size than the tensor statistic {self.name}"
-                    f" {update.size()=}, {self._tensor.size()=}"
-                )
-                self._tensor += update
-            else:
-                assert (
-                    self._shape is None
-                ), "If self._shape is not None, self_.tensor should be initialised in init"
+            update, nb_sample = self._update_function(**kwargs)  # type: ignore
+            assert (self._shape is None or self._shape == update.size()) and (
+                self._tensor is None or self._tensor.size() == update.size()
+            ), (
+                f"The update tensor has a different size than the tensor statistic "
+                f"{self.name} : {self._shape=}, {update.size()=}, "
+                f"{None if self._tensor is None else self._tensor.size()=}"
+            )
+            if self._tensor is None:
                 self._tensor = update
+            else:
+                self._tensor += update
             self.samples += nb_sample
             self.updated = True
+            return update, nb_sample
 
     def init(self):
-        if self._shape is None:
-            self._tensor = None
-        else:
-            self._tensor = torch.zeros(self._shape, device=self.device)
-        self.samples = 0
+        self.reset()
 
     def reset(self):
         self._tensor = None
@@ -110,5 +102,126 @@ class TensorStatistic:
         else:
             assert (
                 self._tensor is not None
-            ), f"If the number of samples is not zero the tensor should not be None."
+            ), "If the number of samples is not zero the tensor should not be None."
             return self._tensor / self.samples
+
+
+class TensorStatiticWithEstimationError(TensorStatistic):
+    """
+    Extends TensorStatistic with estimated quadratic error.
+
+    Extends TensorStatistic to compute an estimation of the quadratic error of the current
+    estimate to the true expectation. This is done by computing the trace of the
+    covariance matrix of the random variable averaged on a batch. The trace is computed
+    incrementally using a stopping criterion based on a relative precision.
+
+    Note that the precision of the trace computation can be controlled by the user, and
+    the true precision of the trace
+    will not be guaranteed to be below this value, indeed if trace_precision is set to
+    eps, the expected relative precision
+    on the trace computation will be of order sqrt(eps).
+
+    Example:
+        We want to compute the average of a set of tensors of shape (2, 3) in data
+        loader `data_loader`. We can use the following code:
+
+            ```python
+            tensor_statistic = TensorStatisticWithEstimationError(
+                update_function=lambda data: (data.sum(dim=0), data.size(0)),
+                name="Average",
+            )
+            for data_batch in data_loader:
+                tensor_statistic.updated = False
+                tensor_statistic.update(data_batch)
+                if tensor_statistic.error() < 0.01:
+                    break
+            print(tensor_statistic())
+            print(tensor_statistic.error())
+            ```
+    """
+
+    def __init__(
+        self,
+        shape: tuple[int, ...] | None,
+        update_function: (
+            Callable[[Any], tuple[torch.Tensor, int]]
+            | Callable[[], tuple[torch.Tensor, int]]
+        ),
+        device: torch.device | str | None = None,
+        name: str | None = None,
+        trace_precision: float = 1e-3,
+    ) -> None:
+        """
+        Initialise the tensor information.
+
+        Parameters
+        ----------
+        shape: tuple[int, ...] | None
+            shape of the tensor to compute, if None use the shape of the first update
+        update_function: Callable[[Any], tuple[torch.Tensor, int]]
+            function to update the tensor and compute the batch covariance
+        name: str
+            used for debugging
+        trace_precision: float
+            relative precision for the trace computation, default 1e-3
+        """
+        super().__init__(shape, update_function, device, name)
+        self._square_norm_sum = 0
+        self.trace_precision = trace_precision
+        # relative precision stopping criterion for the trace computation
+        self._compute_trace = True  # whether to continue computing the trace covariance
+        self._batches = 0
+        self._trace = None  # trace of the covariance matrix
+        # (of the random variable obtain when averaging on a batch)
+
+    def reset(self):
+        super().reset()
+        self._square_norm_sum = 0
+        self._compute_trace = True
+        self._batches = 0
+        self._trace = None
+
+    def error(self) -> float:
+        """
+        Returns an estimation of the quadratic error of the current estimate to the true
+        expectation.
+        If the trace has not been computed accurately enough, NO warning is
+        raised and the error estimate may be inaccurate.
+        If only one batch has been used to compute the statistic, returns infinity.
+
+        Returns
+        -------
+        float
+            estimation of the quadratic error of the current estimate to the true
+            expectation
+        """
+        assert self._trace is not None, "The trace has not been computed yet."
+        if self._batches == 1:
+            return float("inf")
+        return self._trace / self._batches
+
+    @torch.no_grad()
+    def update(self, **kwargs) -> tuple[torch.Tensor, int] | None:
+        if self.updated is False:
+            update, nb_sample = super().update(**kwargs)  # type: ignore (we are sure updated is False here)
+            assert isinstance(
+                self._tensor, torch.Tensor
+            )  # self._tensor should not be None here
+            self._batches += 1
+
+            if self._compute_trace:
+                self._square_norm_sum += update.pow(2).sum().item() / (nb_sample**2)
+                mu_n_norm = self._tensor.pow(2).sum().item() / (self.samples**2)
+                trace_covariance = (self._square_norm_sum) / (self._batches) - mu_n_norm
+                # trace_covariance := Ê[||X||^2] - ||mu_n||^2
+                # where X is the random variable obtained by averaging on a batch
+                # and mu_n is the current estimate of the expectation
+                if self._trace is not None:
+                    delta_trace_covariance = trace_covariance - self._trace
+                    if abs(delta_trace_covariance) < self.trace_precision * abs(
+                        trace_covariance
+                    ):
+                        self._compute_trace = False
+                self._trace = trace_covariance
+
+            return update, nb_sample